extracted the components and related functions into a separate configuration so that we only have one place to maintain. Fixed the readme and cleaned up the grunt files comments that were left over from the backbone stuff
