+++ client/assets/html/index.html
@@ -14,21 +14,16 @@
       <div ui-view="content"></div>
       <div ui-view="login"></div>
     </div>
-    <!-- Shims-->
     <script src="assets/js/modernizr/modernizr.js"></script>
-    <!-- jQuery and Related-->
     <script src="assets/js/jquery/dist/jquery.js"></script>
     <script src="assets/js/select2/select2.js"></script>
     <script src="assets/js/messenger/build/js/messenger.js"></script>
-    <!-- bootstrap-->
     <script src="assets/js/bootstrap/dist/js/bootstrap.js"></script>
-    <!-- AngularJS libraries-->
     <script src="assets/js/angular/angular.js"></script>
     <script src="assets/js/angular-cookies/angular-cookies.js"></script>
     <script src="assets/js/angular-resource/angular-resource.js"></script>
     <script src="assets/js/angular-sanitize/angular-sanitize.js"></script>
     <script src="assets/js/angular-animate/angular-animate.js"></script>
-    <!-- Angular UI libraries-->
     <script src="assets/js/angular-ui-router/release/angular-ui-router.js"></script>
     <script src="assets/js/angular-ui-select2/src/select2.js"></script>
     <script src="assets/js/angular-ui-bootstrap/src/position/position.js"></script>
@@ -36,9 +31,7 @@
     <script src="assets/js/angular-ui-bootstrap/src/datepicker/datepicker.js"></script>
     <script src="assets/js/angular-ui-bootstrap/src/pagination/pagination.js"></script>
     <script src="assets/js/angular-ui-bootstrap/src/buttons/buttons.js"></script>
-    <!--NProgress-->
     <script src="assets/js/nprogress/nprogress.js"></script>
-    <!-- utilities-->
     <script src="assets/js/lodash/dist/lodash.js"></script>
     <script src="assets/js/moment/moment.js"></script>
     <script src="src/app/app.js"></script>
